requirements.txt
heroprotocol==2.55.4.91756
mpyq==0.2.5
pandas==2.2.2
Pillow==10.4.0
sortedcontainers==2.4.0
sv_ttk==2.6.0
tkinter_tooltip==3.1.0
Unidecode==1.3.8
